refactor(proposals): extract general tab into ProposalGeneralTab
The General tab (public/slug URL card, client autocomplete, project and
email-config forms, expiration quick-input, hosting tier previews and
the next-action button) moves verbatim into
components/panel/proposal/ProposalGeneralTab.vue. The page keeps
ownership of the reactive form, hydration, handleUpdate, the effective
totals (the header shows them too) and the payment-percentage sync — the
tab receives them as props and emits
update/toggle-automations/open-email-preview/toggle-active/next-action/
open-actions/client-selected/create-inline-client/
normalize-payment-percentage. DEFAULT_METHOD_PHASES relocates to
proposals_constants.js (shared by hydration and the tab).

edit.vue lands at 1,374 lines (3,205 at the start of this branch); this
completes the B4 split. Verified: 27 E2E (edit, client-autocomplete,
slug, send, update-from-json, discount-offer) + 179 unit tests green.
